feat: add integration sub-command to add command
