+++ packages/cli/src/commands/add/add-integration.ts
@@ -1,0 +1,3 @@
+export type Integration = "ant-design" | "react-router";
+
+export const integrationChoices: Integration[] = ["ant-design", "react-router"];
