+++ packages/cli/src/commands/add/index.ts
@@ -1,6 +1,9 @@
 import { Argument, Command } from "commander";
 import { Provider, createProviders, providerArgs } from "./create-provider";
 import { createResources } from "./create-resource";
+import { Integration, integrationChoices } from "./add-integration";
+import { integrateAntDesign } from "./integrations/ant-design";
+import { integrateReactRouter } from "./integrations/react-router";
 
 const load = (program: Command) => {
     return program
@@ -25,6 +28,15 @@ const load = (program: Command) => {
                     "list,create,edit,show",
                 )
                 .action(createResourceAction),
+        )
+        .addCommand(
+            new Command("integration")
+                .addArgument(
+                    new Argument("<name>", "Name of the integration").choices(
+                        integrationChoices,
+                    ),
+                )
+                .action(addIntegrationAction),
         );
 };
 
@@ -47,6 +59,16 @@ const createResourceAction = async (
         },
         resources,
     );
+};
+
+const addIntegrationAction = async (name: Integration) => {
+    if (name === "ant-design") {
+        await integrateAntDesign();
+    }
+
+    if (name === "react-router") {
+        await integrateReactRouter();
+    }
 };
 
 export default load;
