+++ packages/cli/src/transformers/integrations/ant-design.ts
@@ -6,30 +6,9 @@ import {
     wrapElement,
 } from "../../utils/codeshift";
 import { prettierFormat } from "../../utils/swizzle/prettierFormat";
-import execa from "execa";
 import { API, Collection, FileInfo, JSCodeshift } from "jscodeshift";
 
 export const parser = "tsx";
-
-// runs .bin/jscodeshift with the default export transformer on the current directory
-export const setupAntD = async () => {
-    const jscodeshiftExecutable = require.resolve(".bin/jscodeshift");
-    const { stderr } = execa.sync(jscodeshiftExecutable, [
-        "./",
-        "--extensions=ts,tsx,js,jsx",
-        "--parser=tsx",
-        `--transform=${__dirname}/../src/transformers/setup/antd.ts`,
-        `--ignore-pattern=.cache`,
-        `--ignore-pattern=node_modules`,
-        `--ignore-pattern=build`,
-        `--ignore-pattern=.next`,
-        `--ignore-pattern=dist`,
-    ]);
-
-    if (stderr) {
-        console.log(stderr);
-    }
-};
 
 export default async function transformer(file: FileInfo, api: API) {
     const j = api.jscodeshift;
