+++ packages/cli/src/transformers/integrations/fixtures/ant-design/base-headless.ts
