+++ packages/cli/src/transformers/integrations/fixtures/react-router/with-component-imports.ts
