+++ packages/cli/src/transformers/integrations/fixtures/react-router/with-existing-react-router-setup.ts
