+++ packages/cli/src/transformers/integrations/fixtures/vite-starter.ts
