+++ packages/cli/src/transformers/integrations/react-router.ts
@@ -1,5 +1,4 @@
 import { prettierFormat } from "../../utils/swizzle/prettierFormat";
-import execa from "execa";
 import {
     API,
     Collection,
@@ -16,30 +15,6 @@ import {
 } from "../../utils/codeshift";
 
 export const parser = "tsx";
-
-// runs .bin/jscodeshift with the default export transformer on the current directory
-export const setupReactRouter = async () => {
-    const jscodeshiftExecutable = require.resolve(".bin/jscodeshift");
-    const { stderr, stdout } = execa.sync(jscodeshiftExecutable, [
-        "./",
-        "--extensions=ts,tsx,js,jsx",
-        "--parser=tsx",
-        `--transform=${__dirname}/../src/transformers/setup/react-router.ts`,
-        `--ignore-pattern=.cache`,
-        `--ignore-pattern=node_modules`,
-        `--ignore-pattern=build`,
-        `--ignore-pattern=.next`,
-        `--ignore-pattern=dist`,
-    ]);
-
-    if (stdout) {
-        console.log(stdout);
-    }
-
-    if (stderr) {
-        console.log(stderr);
-    }
-};
 
 export default async function transformer(file: FileInfo, api: API) {
     const j = api.jscodeshift;
